separated plain-caop tests
